+++ lib/generators/happy_seed/admin/templates/app/admin/user.rb
@@ -1,1 +1,4 @@
-ActiveAdmin.register User
+begin
+  ActiveAdmin.register User
+rescue NameError
+end